Fixes some compiler warnings

+++ src/iocore/cache/CacheVC.cc
@@ -88,7 +88,7 @@ DbgCtl dbg_ctl_cache_reenable{"cache_reenable"};
 #define USELESS_REENABLES          // allow them for now
 // #define VERIFY_JTEST_DATA
 
-extern int cache_config_ram_cache_cutoff;
+extern int64_t cache_config_ram_cache_cutoff;
 
 /* Next block with some data in it in this partition.  Returns end of partition if no more
  * locations.
